feat: logging requirements and logging hook (#269)
As discussed in recent meetings, we've had complaints about this from
multiple sources. This PR adds a stipulation that no logging is done by
the SDK during flag evaluation. It also defines a very simple `logging
hook` concept as an included utility. These should be very simple to
write and will provide all the needed logging but give authors more
control than built in log statements. It will also be a very nice intro
to the hooks concept for users.

Here's the Java implementation:
https://github.com/open-feature/java-sdk/pull/1084

---------

Signed-off-by: Todd Baert <todd.baert@dynatrace.com>