fix: Clarify scenario descriptions in context merging feature (#304)
<!-- Please use this template for your pull request. -->
<!-- Please use the sections that you need and delete other sections -->

## This PR
<!-- add the description of the PR here -->

This pull request includes updates to the
`specification/assets/gherkin/contextMerging.feature` file to improve
the clarity of scenario descriptions by specifying the context
(transaction, hook, or both) in which the context entries are added or
overwritten.

Changes to scenario descriptions:

* Updated the scenario description to specify that the context entry is
added for a transaction.
* Updated the scenario description to specify that the context entry is
added for a hook.
* Updated the scenario description to specify that the context entry is
added for both a transaction and a hook.
* Updated the scenario outline description to specify that the context
entry overwrites values for a transaction.
* Updated the scenario outline description to specify that the context
entry overwrites values for a hook.
* Updated the scenario outline description to specify that the context
entry overwrites values for both a transaction and a hook.

### Related Issues
<!-- add here the GitHub issue that this PR resolves if applicable -->

Fixes #303

Signed-off-by: André Silva <2493377+askpt@users.noreply.github.com>