fix: Broaden definition of hook data value types. (#307)
## This PR
Loosens the value definition for hook data.

In the example within the spec, and within the implementation in the
dotnet-sdk, the intent was to allow any type of data. For example
creating an open telemetry span in before, storing it hook data, and
ending it in after.

### Notes
Related to: https://github.com/open-feature/dotnet-sdk/pull/387

---------

Signed-off-by: Ryan Lamb <4955475+kinyoklion@users.noreply.github.com>